Also check for deletion permission. Closes #137

+++ backend/apps/matching/views/staff_approvals.py
@@ -24,7 +24,10 @@ class ApproveParticipantsView(View):
     def get(self, request, p_type):
         # If you are not a staff user, the method decorator takes care of showing a 404 page,
         # so if you are a staff user, we throw a 403 here with an error message
-        if not request.user.has_perm("matching.can_approve_type_%s" % p_type.lower()):
+        if not (
+            request.user.has_perm("matching.can_approve_type_%s" % p_type.lower())
+            or request.user.has_perm("matching.delete_participant%s" % p_type.lower())
+        ):
             raise PermissionDenied("You currently don't have the permission to access this page")
         table_approved = ApproveParticipantTable[p_type](
             Participant[p_type].objects.filter(is_approved=True)
@@ -58,18 +61,27 @@ class ApproveParticipantsView(View):
 
     def post(self, request, *args, **kwargs):
         post_params = self.request.POST
+        p_type = self.kwargs["p_type"]
         if "delete" in post_params:
-            p = get_object_or_404(Participant[self.kwargs["p_type"]], uuid=post_params["uuid"])
+            if not request.user.has_perm("matching.delete_participant%s" % p_type.lower()):
+                raise PermissionDenied(
+                    "You currently don't have the permission to delete %s type of users" % p_type
+                )
+            p = get_object_or_404(Participant[p_type], uuid=post_params["uuid"])
             name = p.user
             p.user.delete()
             text = format_lazy(_("You deleted the participant with email '{name}'."), name=name)
             messages.add_message(self.request, messages.INFO, text)
         elif "change_approval" in post_params:
-            p = get_object_or_404(Participant[self.kwargs["p_type"]], uuid=post_params["uuid"])
+            if not request.user.has_perm("matching.can_approve_type_%s" % p_type.lower()):
+                raise PermissionDenied(
+                    "You currently don't have the permission to approve %s type of users" % p_type
+                )
+            p = get_object_or_404(Participant[p_type], uuid=post_params["uuid"])
             p.change_approval(self.request.user)
             name = p.user
             text = format_lazy(
                 _("You changed the approval of the participant with email '{name}'."), name=name
             )
             messages.add_message(self.request, messages.INFO, text)
-        return self.get(request, self.kwargs["p_type"])
+        return self.get(request, p_type)
